refactor: extract hosted runtime log parser
